requirements.txt
imutils==0.5.2
opencv-python==4.1.0.25
Pillow==6.0.0
pytesseract==0.2.6
aiohttp==3.5.4
async-timeout==3.0.1
attrs==19.1.0
certifi==2019.3.9
chardet==3.0.4
discord.py==1.0.1
idna==2.8
multidict==4.5.2
requests==2.21.0
urllib3==1.24.2
websockets==6.0
yarl==1.3.0

absl-py==0.7.1
astor==0.7.1
atomicwrites==1.3.0
attrs==19.1.0
cycler==0.10.0
decorator==4.4.0
dlib==19.17.0
-e git+https://github.com/thoughtworksarts/EmoPy.git@f1732b8b92ba551cf1ca064581d17ff17d8289ad#egg=EmoPy
gast==0.2.2
graphviz==0.10.1
grpcio==1.20.1
h5py==2.9.0
imageio==2.5.0
imutils==0.5.2
joblib==0.13.2
Keras==2.2.4
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
kiwisolver==1.1.0
Lasagne==0.1
Markdown==3.1
matplotlib==3.0.3
mock==3.0.5
more-itertools==7.0.0
networkx==2.3
numpy==1.14.5
opencv-python==4.1.0.25
Pillow==6.0.0
pluggy==0.11.0
protobuf==3.7.1
py==1.8.0
pydot==1.4.1
pyparsing==2.4.0
pytest==4.5.0
python-dateutil==2.8.0
PyWavelets==1.0.3
PyYAML==5.1
scikit-image==0.15.0
scikit-learn==0.21.1
scikit-neuralnetwork==0.7
scipy==1.2.1
six==1.12.0
tensorboard==1.13.1
tensorflow==1.13.1
tensorflow-estimator==1.13.0
termcolor==1.1.0
Theano==1.0.4
wcwidth==0.1.7
Werkzeug==0.15.4
